+++ frontend/src/pages/GameResultPage/GameResult/FinalButton/FinalButton.utils.ts
