+++ frontend/src/pages/GameResultPage/GameResult/FinalButton/hooks/useResetRoomMutation.ts
@@ -1,0 +1,11 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { resetRoom } from '@/apis/room';
+
+const useResetRoomMutation = (roomId: number) => {
+  return useMutation({
+    mutationFn: async () => await resetRoom(roomId),
+  });
+};
+
+export default useResetRoomMutation;
