+++ frontend/src/pages/GameResultPage/GameResult/GameResult.styled.ts
