+++ frontend/src/pages/GameResultPage/GameResult/GameResultItem/GameResultItem.styled.ts
