+++ frontend/src/pages/GameResultPage/GameResult/GameResultItem/GameResultItem.tsx
