+++ frontend/src/pages/GameResultPage/GameResult/hooks/useScrollControl.ts
