+++ frontend/src/pages/GameResultPage/GameResult/hooks/useScrollState.ts
