+++ frontend/src/pages/GameResultPage/GameResultPage.tsx
@@ -1,6 +1,6 @@
+import GameResult from './GameResult/GameResult';
 import { useIsRoomInitial } from './hooks/useCheckRoomReset';
 
-import GameResult from '@/components/GameResult/GameResult';
 import Content from '@/components/layout/Content/Content';
 
 const GameResultPage = () => {
